refactor(UserService): 트랜잭션 서비스 추상화
