--- app/src/main/java/com/techcourse/service/UserService.java
@@ -1,52 +1,0 @@
-package com.techcourse.service;
-
-import com.techcourse.config.DataSourceConfig;
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
-import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
-import java.util.NoSuchElementException;
-import javax.sql.DataSource;
-import nextstep.jdbc.DataAccessException;
-import org.springframework.jdbc.datasource.DataSourceTransactionManager;
-import org.springframework.transaction.support.DefaultTransactionDefinition;
-
-public class UserService {
-
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
-
-    public UserService(final UserDao userDao, final UserHistoryDao userHistoryDao) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-    }
-
-    public User findById(final long id) {
-        return userDao.findById(id)
-                .orElseThrow(() -> new NoSuchElementException(String.format("id = %d 에 해당하는 유저가 존재하지 않습니다.", id)));
-    }
-
-    public void insert(final User user) {
-        userDao.insert(user);
-    }
-
-    public void changePassword(final long id, final String newPassword, final String createBy) {
-        final var transactionManager = new DataSourceTransactionManager(dataSource());
-        final var transactionStatus = transactionManager.getTransaction(new DefaultTransactionDefinition());
-
-        try {
-            final var user = findById(id);
-            user.changePassword(newPassword);
-            userDao.update(user);
-            userHistoryDao.log(new UserHistory(user, createBy));
-            transactionManager.commit(transactionStatus);
-        } catch (Exception e) {
-            transactionManager.rollback(transactionStatus);
-            throw new DataAccessException(e);
-        }
-    }
-
-    private DataSource dataSource() {
-        return DataSourceConfig.getInstance();
-    }
-}
